requirements.txt
asgiref==3.2.7
Django==3.0.6
django-crispy-forms==1.9.0
django-filter==2.2.0
gunicorn==20.0.4
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.0.1
django_base64field
dlib
django-heroku
numpy
sklearn==0.0
djangorestframework==3.11.0
python-status
opencv-python-headless

